refactor: move login page to [slug]/login
